requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
autopep8==2.3.2
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==5.0.2
Django==5.0
django-allauth==0.61.1
django-cockroachdb==5.0
django-cors-headers==3.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
django-silk==5.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
packaging==24.2
python-decouple==3.8
psycopg2==2.9.10
requests==2.32.3
Pillow==10.4.0
django_jazzmin==3.0.1
django_silk==5.3.1
celery==5.4.0
redis==5.2.1
django-phonenumber-field~=8.0.0
django-filter==25.1.0
phonenumbers==8.12.31
pillow==10.4.0
prompt_toolkit==3.0.50
psycopg2
pycodestyle==2.12.1
PyJWT==2.8.0
pypdf==5.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
cryptography==42.0.5
jsonschema==4.14.0
django-celery-beat==2.8.1
